requirements.txt
# local package
pint~=0.20.1
python-dotenv~=0.21.1
PyYAML~=6.0.1
scipy~=1.10.1
numpy~=1.23.5
setuptools~=69.0.2
pandas~=1.3.5
tqdm~=4.66.4
click~=7.1.2
rainflow~=3.2.0
matplotlib~=3.7.1
rich~=13.6.0
schema~=0.7.5
joblib~=1.3.2
torch~=2.2.2
ply~=3.11
scikit-learn~=1.3.2
seaborn~=0.13.2